Compilation fix
git-svn-id: file:///home/roberto/Git/svn@6813 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 97f975c86413c73c73aeb2856cc42d3538583cb1

+++ trunk/xmipp/libraries/data/normalize.cpp
@@ -546,7 +546,7 @@ void ProgNormalize::show()
 void ProgNormalize::preProcess()
 {
     int Zdim, Ydim, Xdim;
-    size_t Ndim;
+    unsigned long Ndim;
     ImgSize(mdIn, Xdim, Ydim, Zdim, Ndim);
 
     if (!enable_mask)
